requirements.txt
requests
psutil
pystray
Pillow
pyinstaller